prototype(apme): move Git Repos extensions into dynamic plugin packaging
Relocate gitRepositoriesExtensionsApiFactory into backstage-apme for
rhdh-local dynamic plugin wiring; add export-dynamic scripts and backend
dynamic installer entrypoint.
