+++ plugins/backstage-apme-common/package.json
@@ -68,14 +68,14 @@
   },
   "dependencies": {
     "@ansible/backstage-rhaap-common": "workspace:^",
-    "@backstage/backend-plugin-api": "^1.0.0",
-    "@backstage/catalog-model": "^1.7.0",
-    "@backstage/config": "^1.2.0",
-    "@backstage/core-plugin-api": "^1.9.0",
-    "@backstage/errors": "^1.2.4"
+    "@backstage/backend-plugin-api": "^1.8.0",
+    "@backstage/catalog-model": "^1.7.7",
+    "@backstage/config": "^1.3.6",
+    "@backstage/core-plugin-api": "^1.12.4",
+    "@backstage/errors": "^1.2.7"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.27.0",
+    "@backstage/cli": "^0.36.0",
     "typescript": "^5.0.0"
   }
 }
