+++ plugins/backstage-apme/app-config.janus-idp.yaml
@@ -8,16 +8,5 @@ dynamicPlugins:
         - name: apmeIcon
           importName: SecurityIcon
           module: '@material-ui/icons'
-      dynamicRoutes:
-        - path: /apme
-          importName: ApmePage
-          menuItem:
-            icon: apmeIcon
-            text: APME
-        - path: /apme/project/:projectId
-          importName: ApmeProjectPage
-      entityTabs:
-        - path: /apme
-          title: APME
-          mountPoint: entity.page.apme
-          importName: ApmeEntityTab
+      apiFactories:
+        - importName: gitRepositoriesExtensionsApiFactory
