+++ plugins/backstage-apme/package.json
@@ -3,8 +3,13 @@
   "version": "0.1.0",
   "description": "Frontend plugin for APME - Ansible Policy & Modernization Engine",
   "backstage": {
-    "role": "frontend-plugin"
+    "role": "frontend-plugin",
+    "pluginId": "apme",
+    "pluginPackages": [
+      "@ansible/plugin-backstage-apme"
+    ]
   },
+  "sideEffects": false,
   "publishConfig": {
     "access": "public"
   },
@@ -12,7 +17,9 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "files": [
-    "dist"
+    "dist",
+    "config.d.ts",
+    "dist-scalprum"
   ],
   "configSchema": "config.d.ts",
   "scripts": {
@@ -22,7 +29,8 @@
     "test": "backstage-cli package test",
     "prepack": "backstage-cli package prepack",
     "postpack": "backstage-cli package postpack",
-    "start": "backstage-cli package start"
+    "start": "backstage-cli package start",
+    "export-dynamic": "rhdh-cli plugin export --embed-package @ansible/backstage-apme-common --embed-package @ansible/backstage-rhaap-common"
   },
   "dependencies": {
     "@ansible/backstage-apme-common": "workspace:^",
