+++ plugins/backstage-apme/src/utils/fetchAllProjectViolations.ts
@@ -14,5 +14,17 @@
  * limitations under the License.
  */
 
-export { ApmeViolationsTable } from './ApmeViolationsTable';
-export type { ApmeViolationsTableProps } from './ApmeViolationsTable';
+import type { Violation } from '@ansible/backstage-apme-common/types';
+import type { ApmeApi } from '../api';
+
+const MAX_VIOLATIONS = 5000;
+
+/** Fetch gateway violations up to the API limit (no server-side offset yet). */
+export async function fetchAllProjectViolations(
+  apmeApi: ApmeApi,
+  projectId: string,
+  totalHint?: number,
+): Promise<Violation[]> {
+  const limit = Math.min(totalHint ?? MAX_VIOLATIONS, MAX_VIOLATIONS);
+  return apmeApi.getViolations(projectId, { limit });
+}
