+++ plugins/catalog-backend-module-apme/src/dynamic/index.ts
@@ -1,0 +1,8 @@
+import { BackendDynamicPluginInstaller } from '@backstage/backend-dynamic-feature-service';
+
+import catalogModuleApme from '..';
+
+export const dynamicPluginInstaller: BackendDynamicPluginInstaller = {
+  kind: 'new',
+  install: () => catalogModuleApme,
+};
